refactor(tests): share waitForHttpStatus polling helper
Move the helper from snapshot.test.ts into tests/setup.ts, return the
matched response body, and use it in the httpsPorts network test.
